LayoutTests/fast/events/touch/touch-target-expected.txt

Tests that the target of touches match the element where the event originated, not where the touch is currently occurring.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS event.type is "touchstart"
PASS event.touches[0].target.id is "targetA"
PASS event.touches[1].target.id is "targetB"
PASS event.type is "touchmove"
PASS event.touches[0].target.id is "targetA"
PASS event.touches[1].target.id is "targetB"
PASS event.type is "touchmove"
PASS event.touches[0].target.id is "targetA"
PASS event.touches[1].target.id is "targetB"
PASS event.type is "touchstart"
PASS event.touches[0].target.id is "targetB"
PASS event.touches[1].target.id is "targetA"
PASS event.type is "touchmove"
PASS event.touches[0].target.id is "targetB"
PASS event.touches[1].target.id is "targetA"
PASS successfullyParsed is true

TEST COMPLETE
